Move object/icecrusher.cpp&hpp to badguy/icecrusher.cpp&hpp
