+++ src/badguy/icecrusher.cpp
@@ -15,7 +15,7 @@
 //  You should have received a copy of the GNU General Public License
 //  along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
-#include "object/icecrusher.hpp"
+#include "badguy/icecrusher.hpp"
 
 #include <algorithm>
 #include <math.h>
