+++ src/badguy/icecrusher.hpp
